moved sensors to subpackage
