+++ cms_perf/sensors/net_load.py
@@ -5,7 +5,7 @@ import enum
 
 import psutil
 
-from .cli_parser import cli_domain, cli_call
+from ..cli_parser import cli_domain, cli_call
 
 
 @cli_domain(name="KIND")
